refactor: server definition and move CLI entry point
Reorganized the server implementation to simplify usage and maintenance. Moved the CLI logic to a dedicated `cli.py` module and updated the setup configuration accordingly. Streamlined the application structure and improved separation of concerns.
